update: schema.sql

+++ ecommerce-batch/src/main/java/fastcampus/ecommerce/batch/domain/product/report/ManufacturerReport.java
@@ -15,7 +15,7 @@ public class ManufacturerReport {
     // 배치가 실행된 날짜
     private LocalDate statDate = LocalDate.now();
     private String manufacturer;
-    private Integer productCount;
+    private Long productCount;
     private BigDecimal avgSalesPrice;
-    private Integer potentialSalesAmount;
+    private BigDecimal potentialSalesAmount;
 }
