+++ sql/schema.sql
@@ -75,7 +75,7 @@ create table brand_reports
 (
     stat_date              date           not null,
     brand                  varchar(255)   not null,
-    product_count          integer        not null,
+    product_count          bigint         not null,
     avg_sales_price        decimal(15, 0) not null,
     max_sales_price        decimal(15, 0) not null,
     min_sales_price        decimal(15, 0) not null,
@@ -89,7 +89,7 @@ create table category_reports
 (
     stat_date              date           not null,
     category               varchar(255)   not null,
-    product_count          integer        not null,
+    product_count          bigint         not null,
     avg_sales_price        decimal(15, 0) not null,
     max_sales_price        decimal(15, 0) not null,
     min_sales_price        decimal(15, 0) not null,
@@ -102,7 +102,7 @@ create table manufacturer_reports
 (
     stat_date              date           not null,
     manufacturer           varchar(255)   not null,
-    product_count          integer        not null,
+    product_count          bigint         not null,
     avg_sales_price        decimal(15, 0) not null,
     potential_sales_amount decimal(20, 0) not null,
     primary key (stat_date, manufacturer)
@@ -112,7 +112,7 @@ create table product_status_reports
 (
     stat_date          date           not null,
     product_status     varchar(255)   not null,
-    product_count      integer        not null,
+    product_count      bigint         not null,
     avg_stock_quantity decimal(15, 0) not null,
     primary key (stat_date, product_status)
 );
